+++ changelog.md
@@ -4,6 +4,7 @@
 
 ## Features
 - Add `--cut-out` and `--add-in`
+- Set the temporary directory with `--temp-dr`
 - Support export aliases (`-exp`, `-exs`, etc.)
 - Opening the new file now works on Windows and Linux.